requirements.txt
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.3
geographiclib==2.1
geopy==2.4.1
gunicorn==25.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==26.0
pillow==12.1.1
python-dotenv==1.2.2
qrcode==8.2
Werkzeug==3.1.7
